requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
Faker==37.1.0
gunicorn==23.0.0
packaging==24.2
pillow==11.2.1
psycopg2==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
